refactor(component): rename to align with content
